modularization of the code base
