+++ .plan
@@ -96,6 +96,11 @@ Grouped by sections and then sorted by priority (more or less).
 * [_] completely switch to JS modules at some point
 * [_] use more generator functions when possible
 
+#### Event Bus Todo List
+
+* [_] SV and Host: listening to cvar changes
+* [_] Draw: listen to filesystem being/end disc
+
 ### Code smell
 
 * [_] SV: cleanup the `SV.WriteEntitiesToClient` mess a bit, make it more flexible
@@ -321,9 +326,24 @@ class MainMenu extends BaseMenu {
 
 ## Log
 
+### 2025-07-14
+
+* tightened jsconfig and eslint config a bit to enforce a stricter standard, it already detected some silent bugs
+
 ### 2025-07-12
 
 * I asked Claude Sonnet 4 to improve the Key input mess (especially when using a non-us keyboard layout)
+* I started to refactor the engine code:
+  * completely moving to ESM
+  * throwing out unused code
+  * breaking larger files into smaller better managable ones
+  * introduced a registry to hold references to singletons (workaround to avoid circular dependencies)
+* next steps I’m planning on the engine side:
+  * decouple tightly decoupled components
+  * implement an eventbus that can be accessed by the regular code and worker
+  * reduce reliance on the registry to an absolute minimum
+  * make stuff async as much as possible
+    * uploading textures and building VBA in workers
 
 ### 2025-07-09
