+++ quakeshack-dedicated.mjs
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import EngineLauncher from './source/engine/main-dedicated.mjs';
+
+await EngineLauncher.Launch();
