--- source/Draw_node.js
@@ -1,6 +1,0 @@
-/* global Draw */
-
-// eslint-disable-next-line no-global-assign
-Draw = {};
-Draw.BeginDisc = function() {};
-Draw.EndDisc = function() {};
