+++ source/engine/main-browser.mjs
@@ -1,0 +1,9 @@
+import { registry } from './registry.mjs';
+
+export default class EngineLauncher {
+  static async Launch() {
+    console.log('Launching engine in browser mode...');
+
+    registry.isDedicatedServer = false;
+  }
+};
